+++ Semana_4/node/src/config/upload.ts
@@ -3,9 +3,11 @@ import path from "path";
 import crypto from "crypto";
 
 const tmpFolder = path.resolve(__dirname, "..", "..", "tmp")
+const upFolder = path.resolve(__dirname, "..", "..", "tmp",'uploads')
 
 export default {
   directory : tmpFolder,
+  uploadDirectory : upFolder,
   storage: multer.diskStorage({
     destination: tmpFolder,
     filename(request, file, callback) {